Refactor: Extract CameraViewComponent and MicrophoneComponent
- I extracted camera view functionality into a new CameraViewComponent.
  - This component handles video feed, image capture, file upload for images, and interaction with the Gemini API for image-based queries.
  - HomeComponent now delegates camera operations to CameraViewComponent.

- I extracted microphone listening UI and toggle functionality into a new MicrophoneComponent.
  - This component displays listening status and provides a button to toggle listening.
  - HomeComponent now uses MicrophoneComponent to manage listening state and UI.

- Both new components are non-standalone and have been added to AppModule.
- I confirmed that FormsModule is available for CameraViewComponent's use of ngModel.
- I ensured the application builds successfully after these changes.

This refactoring improves modularity and separation of concerns within the application.
