+++ src/app/app.module.ts
@@ -6,6 +6,8 @@ import { AppRoutingModule } from './app-routing.module';
 import {RootComponent} from './components/root/root.component';
 import { HomeComponent } from './pages/home/home.component';
 import {LayoutComponent} from './components/layout/layout.component';
+import { CameraViewComponent } from './components/camera-view/camera-view.component';
+import { MicrophoneComponent } from './components/microphone/microphone.component';
 
 @NgModule({
   declarations: [
@@ -13,7 +15,9 @@ import {LayoutComponent} from './components/layout/layout.component';
     RootComponent,
 
 
-    HomeComponent
+    HomeComponent,
+    CameraViewComponent,
+    MicrophoneComponent
   ],
   imports: [
     BrowserModule,
