+++ src/app/components/camera-view/camera-view.component.scss
@@ -1,0 +1,2 @@
+/* Styles for CameraViewComponent will go here */
+/* For now, it's empty as per the plan */
